Now abusing the null byte to expand pages' root paths!  Yay!
From this commit on, a null byte (\x00) followed by a path and another
null byte will be replaced with the root path for the page given between
the two null bytes, or for the current page if the path is empty.  This
replacement is done in a new `rootify` Liquid tag, and the `root` tag
now outputs the proper null-byte-laden sequence.

This is absolutely fucking ugly as pig shit, but it's necessary in order
to get relative roots working in places like blog post listings.  (The
`<base>` HTML tag only allows relative values for `href` starting in
HTML 5, so I can't expect that to work in older browsers.)

If you want to include a null byte in a page or post for whatever
reason, use `&#x00;` in your HTML or Markdown code.