[plugins/root] modify the null-byte abuse protocol to support options, e.g. absolue vs. relative root
The null-byte abuse protocol now has a third null byte.  The text in
between the second and third null byte is an options field, which
may be empty.  Currently, the options field is just a character "a"
or "r":  "a" means absolute root, and "r" means relative root.  If
the option is omitted, then the default is used.