Add support for Material icons
This uses the Material Icons webfont
(<https://google.github.io/material-design-icons/#icon-font-for-the-web>),
which uses ligatures to allow one to just type out the name of the
desired icon instead of having to hunt down an image or manually specify
a codepoint.

To use a Material icon in text, simply type out the icon name inside an
element with the `.material-icons` class.  Multiple words in a single
icon name are separated with underscores.

To use a Material icon as a page's icon, use `material:<icon name>` as
the value for the `icon` property.  Multiple words in a single icon name
may be separated with underscores, hyphens, or spaces.  There is
currently no support for using Material icons as the page's favicon;
this currently only affects navigation elements.  Favicon support may be
implemented later by dynamically generating them in JavaScript using the
Canvas element.